zerodha_bot/requirements.txt
kiteconnect
pandas
numpy
yfinance
scikit-learn
telethon
python-dotenv
pytz
openpyxl
